Major patch to resources.
