--- src/main/java/com/veteam/voluminousenergy/recipe/RecipeConstants.java
@@ -1,30 +1,0 @@
-package com.veteam.voluminousenergy.recipe;
-
-import com.veteam.voluminousenergy.VoluminousEnergy;
-import net.minecraft.util.ResourceLocation;
-
-public class RecipeConstants {
-
-        /*
-    public static final RegistryObject<IRecipeSerializer<?>> PRIMITIVE_BLAST_FURNACE_RECIPE = VE_RECIPES.register("primitive_blast_furnacing", () -> PrimitiveBlastFurnaceRecipe.serializer);
-    public static final RegistryObject<IRecipeSerializer<?>> CRUSHER_RECIPE = VE_RECIPES.register("crushing", () -> CrusherRecipe.serializer);
-    public static final RegistryObject<IRecipeSerializer<?>> ELECTROLYZER_RECIPE = VE_RECIPES.register("electrolyzing", () -> ElectrolyzerRecipe.serializer);
-    public static final RegistryObject<IRecipeSerializer<?>> CENTRIFUGAL_AGITATOR_RECIPE = VE_RECIPES.register("centrifugal_agitating", () -> CentrifugalAgitatorRecipe.SERIALIZER);
-    public static final RegistryObject<IRecipeSerializer<?>> COMPRESSOR_RECIPE = VE_RECIPES.register("compressing", () -> CompressorRecipe.SERIALIZER);
-    public static final RegistryObject<IRecipeSerializer<?>> STIRLING_GENERATOR_RECIPE = VE_RECIPES.register("stirling", () -> StirlingGeneratorRecipe.SERIALIZER);
-    public static final RegistryObject<IRecipeSerializer<?>> COMBUSTION_GENERATOR_OXIDIZER_RECIPE = VE_RECIPES.register("stirling", () -> CombustionGeneratorOxidizerRecipe.SERIALIZER);
-    public static final RegistryObject<IRecipeSerializer<?>> COMBUSTION_GENERATOR_FUEL_RECIPE = VE_RECIPES.register("fuel_combustion", () -> CombustionGeneratorFuelRecipe.SERIALIZER);
-    public static final RegistryObject<IRecipeSerializer<?>> AQUEOULIZER_RECIPE = VE_RECIPES.register("aqueoulizing", () -> AqueoulizerRecipe.SERIALIZER);
-     */
-
-    public static final ResourceLocation PRIMITIVE_BLAST_FURNACING = new ResourceLocation(VoluminousEnergy.MODID, "primitive_blast_furnacing");
-    public static final ResourceLocation CRUSHING = new ResourceLocation(VoluminousEnergy.MODID, "crushing");
-    public static final ResourceLocation ELECTROLYZING = new ResourceLocation(VoluminousEnergy.MODID, "electrolyzing");
-    public static final ResourceLocation CENTRIFUGAL_AGITATING = new ResourceLocation(VoluminousEnergy.MODID, "centrifugal_agitating");
-    public static final ResourceLocation COMPRESSING = new ResourceLocation(VoluminousEnergy.MODID, "compressing");
-    public static final ResourceLocation STIRLING = new ResourceLocation(VoluminousEnergy.MODID, "stirling");
-    public static final ResourceLocation FUEL_COMBUSTION = new ResourceLocation(VoluminousEnergy.MODID, "fuel_combustion");
-    public static final ResourceLocation OXIDIZING = new ResourceLocation(VoluminousEnergy.MODID, "oxidizer_combustion");
-    public static final ResourceLocation AQUEOULIZING = new ResourceLocation(VoluminousEnergy.MODID, "aqueoulizing");
-
-}
