+++ src/main/java/com/veteam/voluminousenergy/util/RecipeConstants.java
@@ -1,0 +1,18 @@
+package com.veteam.voluminousenergy.util;
+
+import com.veteam.voluminousenergy.VoluminousEnergy;
+import net.minecraft.util.ResourceLocation;
+
+public class RecipeConstants {
+
+    public static final ResourceLocation PRIMITIVE_BLAST_FURNACING = new ResourceLocation(VoluminousEnergy.MODID, "primitive_blast_furnacing");
+    public static final ResourceLocation CRUSHING = new ResourceLocation(VoluminousEnergy.MODID, "crushing");
+    public static final ResourceLocation ELECTROLYZING = new ResourceLocation(VoluminousEnergy.MODID, "electrolyzing");
+    public static final ResourceLocation CENTRIFUGAL_AGITATING = new ResourceLocation(VoluminousEnergy.MODID, "centrifugal_agitating");
+    public static final ResourceLocation COMPRESSING = new ResourceLocation(VoluminousEnergy.MODID, "compressing");
+    public static final ResourceLocation STIRLING = new ResourceLocation(VoluminousEnergy.MODID, "stirling");
+    public static final ResourceLocation FUEL_COMBUSTION = new ResourceLocation(VoluminousEnergy.MODID, "fuel_combustion");
+    public static final ResourceLocation OXIDIZING = new ResourceLocation(VoluminousEnergy.MODID, "oxidizer_combustion");
+    public static final ResourceLocation AQUEOULIZING = new ResourceLocation(VoluminousEnergy.MODID, "aqueoulizing");
+
+}
