+++ src/main/resources/META-INF/mods.toml
@@ -13,7 +13,7 @@ version="1.15.2-0.0.0.0" #mandatory
  # A display name for the mod
 displayName="Voluminous Energy" #mandatory
 # A URL to query for updates for this mod. See the JSON update specification <here>
-updateJSONURL="https://github.com/MikeTheShadow/VoluminousEnergy" #optional
+#updateJSONURL="https://github.com/MikeTheShadow/VoluminousEnergy" #optional
 # A URL for the "homepage" for this mod, displayed in the mod UI
 displayURL="https://github.com/MikeTheShadow/VoluminousEnergy" #optional
 # A file name (in the root of the mod JAR) containing a logo for display
