--- src/main/resources/data/voluminousenergy/recipes/crushing/ores/non-native/aluminum.json
@@ -1,17 +1,0 @@
-{
-  "type": "voluminousenergy:crushing",
-  "ingredient": {
-    "tag": "forge:ores/aluminum",
-    "count": 1
-  },
-  "processTime": 200,
-  "result": {
-    "item": "voluminousenergy:aluminum_dust",
-    "count": 2
-  },
-  "rng": {
-    "item": "voluminousenergy:bauxite_dust",
-    "count": 1,
-    "chance": 0.25
-  }
-}
