--- src/main/resources/data/voluminousenergy/recipes/crushing/ores/non-native/silver.json
@@ -1,17 +1,0 @@
-{
-  "type": "voluminousenergy:crushing",
-  "ingredient": {
-    "tag": "forge:ores/silver",
-    "count": 1
-  },
-  "processTime": 200,
-  "result": {
-    "item": "voluminousenergy:silver_dust",
-    "count": 2
-  },
-  "rng": {
-    "item": "voluminousenergy:galena_dust",
-    "count": 1,
-    "chance": 0.25
-  }
-}
