+++ src/main/resources/data/voluminousenergy/recipes/waffles/compressed_air.json
@@ -1,6 +1,6 @@
 {
   "type": "voluminousenergy:oxidizer_combustion",
-  "processTime": 1600,
+  "process_time": 1600,
   "ingredient": {
     "item": "voluminousenergy:compressed_air_bucket"
   }
